also use cluster in the scheduler
